Make protocol arg to "canonicalize a port" optional. (Fixes #90)
The "canonicalize a port" algorithm is used for both pattern encoding callbacks and for
normalizing URL type inputs to exec() and test().  In the test() and exec() cases we want
the default port to be normalized to empty string, so a protocol value is passed in.  When
used as an encoding callback for pattern parsing, however, we don't get the protocol
and we don't do default port normalization here.  Instead it is done in a separate step
in the URLPattern contructor.