Escape some suffixes trailing `:foo` groups. (#145)
This fixes one of the cases outlined in #145.  When a named group, like
`:foo`, has a suffix that begins with a valid name character, like `bar`,
we run the risk of generating a string that adds the suffix to the name
itself and not as a suffix to the group.  To avoid this we insert an
escape character before the suffix.  For example, we want to generate
`:foo\bar` and not `:foobar`.

This spec change corresponds to this implementation change:

https://chromium-review.googlesource.com/c/chromium/src/+/3318605