Emit `{}` when plain text follows a `:foo` group. (#145)
This fixes another issue highlighted in #145.

In this case we need to detect when fixed text following a `:foo` group
could be mistaken as part of the name.  When this happens we should emit
`{ }` braces around the group to isolate it.  For example, `{:foo}bar`
instead of of `:foobar`.

This spec change corresponds to the following implementation changes:

https://chromium-review.googlesource.com/c/chromium/src/+/3321520
https://chromium-review.googlesource.com/c/chromium/src/+/3328146