URLPattern: Emit '{ }' when there is a preceding non-prefix '/'. (#145)
This commit fixes another issue from #145.

In this case we want to make sure that a character that can be treated
as an implicit group prefix, e.g. `/`, is not incorrectly converted into
a prefix.  For example, if the original string is `/{:foo?}` we don't
want to generate `/:foo?` since that would make the `/` optional.

This CL also includes an additional fix that permits emitting `*` for
wildcards when there is a prefix.  Previously we were being too
conservative in some situations.  For example, we can emit `*/*` instead
of `*/(.*)`.

This commit corresponds to this implementation change:

https://chromium-review.googlesource.com/c/chromium/src/+/3378528