Remove erroneous prefix check when generating string. (#145)
This fixes another issue with generating canonical pattern strings as described
in #145.  In this case the prefix check incorrectly prevented us from
protecting a named group from following character if there was an implicit
prefix.  For example, for a pathname `/:foo\\bar` we would incorrectly produce
`/:foobar` instead of `{/:foo}bar`.