feat: support for case-insensitive matches (#168)
This commit updates the specification document with the support of
case-insensitive matches by introducing a third argument in the
`URLPattern` constructor. This additional document contains a single
boolean property named `ignoreCase`, which can be set to `true` to
enable case-insensitive matches. By default this flag is set to `false`,
thus enabling case-sensitive matches.