Add a section on other specs integrating with URLPattern
A section which covers how other specifications should use URLPattern
and how developer-facing APIs should work is added, along with helpful
algorithms.

One of these is whether a pattern has regexp groups (which may require
an ECMAScript regexp engine); a corresponding WebIDL attribute is added
to expose this property to authors as well.