Do not initialize a URLPatternInit member to null in "initialize a URLPattern"
It is not valid for the baseURL dictionary member to be null, only
either absent or a USVString. Instead, this should be omitted from the
dictionary altogether if no string was provided to this algorithm.

Any string which is invalid will fail later on, when it is to be parsed
as a URL.

Fixes #204.