Make test() and exec() accept URL
The original specification did not explicitly accept a URL object as an
input, and expected implicit conversion from the URL object to the
URLPatternInit object for matching.

This change make the URL object accepted as an input explicitly.
This is a step to fix https://github.com/whatwg/urlpattern/issues/218.