Use the basic URL parser when parsing URLs
Use the basic URL parser when parsing URLs

Blob handling is not required, since the resulting URL is not stored in
any way that would make the blob handling visible. This is consistent
with what the URL constructor and similar uses do.

This change should not be observable.

Partially addresses #242.